+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# Unreleased
+- Add experimental static code diagnostics prefer-trailing-comma-for-collection
+
 # 1.5.1
 - Improve code diagnostics double-literal-format, no-boolean-literal-compare
 - Add experimental static code diagnostics newline-before-return, no-empty-block, avoid-preserve-whitespace-false